+++ hyperspy/tests/drawing/test_plot_signal1d.py
@@ -391,14 +391,7 @@ def test_plot_spectra_linestyle(linestyle):
 
 
 def test_plot_spectra_linestyle_error():
-    from hyperspy.exceptions import VisibleDeprecationWarning
     s = hs.signals.Signal1D(np.arange(100).reshape(2, 50))
-    with pytest.warns(VisibleDeprecationWarning):
-        hs.plot.plot_spectra(s, line_style='--')
-
-    with pytest.raises(ValueError):
-        with pytest.warns(VisibleDeprecationWarning):
-            hs.plot.plot_spectra(s, linestyle='-', line_style='--')
 
     with pytest.raises(ValueError):
         hs.plot.plot_spectra(s, linestyle='invalid')