+++ CHANGES.rst
@@ -134,6 +134,7 @@ Data Visualisation
 - The :py:meth:`~._signals.signal2d.Signal2D.plot` keyword argument
   ``saturated_pixels`` is removed, use ``vmin`` and/or ``vmax`` instead.
 - The ``get_complex`` property of :py:class:`~.drawing.signal1d.Signal1DLine` is removed.
+- The keyword argument ``line_style`` of :py:func:`~.drawing.utils.plot_spectra` have been renamed to ``linestyle``.
 
 
 .. towncrier release notes start